requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23

# Optional for production
gunicorn==21.2.0
psycopg2-binary==2.9.9  # If using PostgreSQL

# Development tools
Flask-DebugToolbar==0.14.1
pytest==7.4.3
pytest-flask==1.3.0